Slave comms

+++ lib/slave/slave.h
@@ -1,7 +1,7 @@
 #pragma once
 
 void slaveSetup();
-void servoAngle(int angle);
+void servoAngle(float angle);
 void motorSpeed(int speed);
 int getEncoders();
 void slaveProcessSerial();
